Removing unused members and moving configuration types up to their folder
