+++ src/AutoMapper/IProfileConfiguration.cs
@@ -17,8 +17,6 @@ namespace AutoMapper
         bool ConstructorMappingEnabled { get; }
         bool AllowNullDestinationValues { get; }
         bool AllowNullCollections { get; }
-        //INamingConvention SourceMemberNamingConvention { get; }
-        //INamingConvention DestinationMemberNamingConvention { get; }
         bool CreateMissingTypeMaps { get; }
 
         IMemberConfiguration DefaultMemberConfig { get; }
@@ -42,9 +40,33 @@ namespace AutoMapper
         string ProfileName { get; }
         IEnumerable<string> GlobalIgnores { get; }
 
+        /// <summary>
+        /// Registers all defined type maps
+        /// </summary>
+        /// <param name="typeMapRegistry">Type map registry</param>
         void Register(TypeMapRegistry typeMapRegistry);
+
+        /// <summary>
+        /// Configured all defined type maps
+        /// </summary>
+        /// <param name="typeMapRegistry">Type map registry</param>
         void Configure(TypeMapRegistry typeMapRegistry);
+
+        /// <summary>
+        /// Created and configures a type map based on conventions if matching
+        /// </summary>
+        /// <param name="typeMapRegistry">Type map registry</param>
+        /// <param name="conventionTypes">Types to match</param>
+        /// <returns>Configured type map, or null if not a match</returns>
         TypeMap ConfigureConventionTypeMap(TypeMapRegistry typeMapRegistry, TypePair conventionTypes);
+
+        /// <summary>
+        /// Creates and configures a closed generic type map based on conventions if matching
+        /// </summary>
+        /// <param name="typeMapRegistry">Type map registry</param>
+        /// <param name="closedTypes">Closed types to create</param>
+        /// <param name="openTypes">Open types to match</param>
+        /// <returns>Configured type map, or null if not a match</returns>
         TypeMap ConfigureClosedGenericTypeMap(TypeMapRegistry typeMapRegistry, TypePair closedTypes, TypePair openTypes);
 	}
 }
