--- src/AutoMapper/ITypeMapFactory.cs
@@ -1,9 +1,0 @@
-namespace AutoMapper
-{
-    using System;
-
-    public interface ITypeMapFactory
-    {
-        TypeMap CreateTypeMap(Type sourceType, Type destinationType, IProfileConfiguration mappingOptions, MemberList memberList);
-    }
-}
